user_interface/requirements.txt
streamlit==1.29.0
streamlit-authenticator==0.2.3
pandas==2.1.3
numpy==1.26.2
plotly==5.18.0
plotly-express==0.4.1
SQLAlchemy==2.0.23
DateTime
pycountry==23.12.11
cleanco==2.2
python-Levenshtein==0.23.0
python-lei==0.1.1

